update old intmax_t params to int
gofmt seems to have made some space-related changes here too...
not quite sure why.
This commit means ui now compiles against the current latest
libui (although I've excluded any updated library binaries from this
commit)

+++ combobox.go
@@ -22,10 +22,10 @@ var comboboxes = make(map[*C.uiCombobox]*Combobox)
 // Combobox also has an entry field that the user can type an alternate
 // choice into.
 type Combobox struct {
-	co	*C.uiControl
-	c	*C.uiCombobox
+	co *C.uiControl
+	c  *C.uiCombobox
 
-	onSelected		func(*Combobox)
+	onSelected func(*Combobox)
 }
 
 // NewCombobox creates a new Combobox.
@@ -41,6 +41,7 @@ func NewCombobox() *Combobox {
 
 	return c
 }
+
 /*TODO
 // NewEditableCombobox creates a new editable Combobox.
 func NewEditableCombobox() *Combobox {
@@ -114,7 +115,7 @@ func (c *Combobox) Selected() int {
 // SetChecked sets the currently select item in the Combobox
 // to index. If index is -1 no item will be selected.
 func (c *Combobox) SetSelected(index int) {
-	C.uiComboboxSetSelected(c.c, C.intmax_t(index))
+	C.uiComboboxSetSelected(c.c, C.int(index))
 }
 
 // OnSelected registers f to be run when the user selects an item in
